refactor: split schema form as separated component (#2068)
